1) sparse_convolution_layer - core functionality added, 2) softmax accuracy improved, 3) weight_decay not applied to biases

+++ nnforge/training_task_state.h
@@ -36,7 +36,7 @@ namespace nnforge
 
 		unsigned int get_current_epoch() const
 		{
-			return history.size() + initial_epoch;
+			return static_cast<unsigned int>(history.size() + initial_epoch);
 		}
 	};
 }
